+++ src/excmds.ts
@@ -70,7 +70,7 @@
 
 // Shared
 import * as Messaging from "@src/lib/messaging"
-import { browserBg, activeTab, activeTabId, activeTabContainerId, openInNewTab, openInNewWindow, openInTab } from "@src/lib/webext"
+import { browserBg, activeTab, activeTabId, activeTabContainerId, openInNewTab, openInNewWindow, openInTab, ABOUT_WHITELIST } from "@src/lib/webext"
 import * as Container from "@src/lib/containers"
 import state from "@src/state"
 import { contentState, ModeName } from "@src/content/state_content"
@@ -78,7 +78,6 @@ import * as UrlUtil from "@src/lib/url_util"
 import * as config from "@src/lib/config"
 import * as aliases from "@src/lib/aliases"
 import * as Logging from "@src/lib/logging"
-import { AutoContain } from "@src/lib/autocontainers"
 import * as CSS from "css"
 import * as Perf from "@src/perf"
 import * as Metadata from "@src/.metadata.generated"
@@ -110,6 +109,13 @@ const logger = new Logging.Logger("excmds")
 
 /** @hidden **/
 const TRI_VERSION = "REPLACE_ME_WITH_THE_VERSION_USING_SED"
+
+// Import excmd libs
+import excmd_rss from "@src/excmds/rss"
+import excmd_fillcmdline from "@src/excmds/fillcmdline"
+import excmd_clipboard from "@src/excmds/clipboard"
+import excmd_open from "@src/excmds/open"
+import excmd_tabs from "@src/excmds/tabs"
 
 //#content_helper
 // {
@@ -170,33 +176,6 @@ export async function getNativeVersion(): Promise<void> {
     Native.getNativeMessengerVersion()
 }
 
-/** @hidden
- * This function is used by rssexec and rssexec completions.
- */
-//#content
-export async function getRssLinks(): Promise<Array<{ type: string; url: string; title: string }>> {
-    const seen = new Set<string>()
-    return Array.from(document.querySelectorAll("a, link[rel='alternate']"))
-        .filter((e: any) => typeof e.href === "string")
-        .reduce((acc, e: any) => {
-            let type = ""
-            // Start by detecting type because url doesn't necessarily contain the words "rss" or "atom"
-            if (e.type) {
-                // if type doesn't match either rss or atom, don't include link
-                if (e.type.indexOf("rss") < 0 && e.type.indexOf("atom") < 0) return acc
-                type = e.type
-            } else {
-                // Making sure that we match either a dot or "xml" because "urss" and "atom" are actual words
-                if (e.href.match(/(\.rss)|(rss\.xml)/i)) type = "application/rss+xml"
-                else if (e.href.match(/(\.atom)|(atom\.xml)/i)) type = "application/atom+xml"
-                else return acc
-            }
-            if (seen.has(e.href)) return acc
-            seen.add(e.href)
-            return acc.concat({ type, url: e.href, title: e.title || e.innerText } as { type: string; url: string; title: string })
-        }, [])
-}
-
 /**
  * Execute [[rsscmd]] for an rss link.
  *
@@ -208,32 +187,7 @@ export async function getRssLinks(): Promise<Array<{ type: string; url: string; 
  */
 //#content
 export async function rssexec(url: string, type?: string, ...title: string[]) {
-    if (!url || url === "") {
-        const links = await getRssLinks()
-        switch (links.length) {
-            case 0:
-                throw new Error("No rss link found on this page.")
-                break
-            case 1:
-                url = links[0].url
-                title = [links[0].title]
-                type = links[0].type
-                break
-            default:
-                return fillcmdline("rssexec")
-        }
-    }
-    let rsscmd = config.get("rsscmd")
-    if (rsscmd.match("%[uty]")) {
-        rsscmd = rsscmd
-            .replace("%u", url)
-            .replace("%t", title.join(" "))
-            .replace("%y", type || "")
-    } else {
-        rsscmd += " " + url
-    }
-    // Need actual excmd parsing here.
-    return controller.acceptExCmd(rsscmd)
+    return excmd_rss.rssexec(url, type, ...title)
 }
 
 /**
@@ -524,81 +478,7 @@ export async function fixamo() {
  */
 //#background
 export async function nativeopen(...args: string[]) {
-    const index = args.findIndex(arg => !arg.startsWith("-"))
-    let firefoxArgs = []
-    if (index >= 0) {
-        firefoxArgs = args.slice(0, index)
-    }
-    const url = args.slice(firefoxArgs.length).join(" ")
-
-    if (await Native.nativegate()) {
-        // First compute where the tab should be
-        const pos = await config.getAsync("tabopenpos")
-        let index = (await activeTab()).index + 1
-        switch (pos) {
-            case "last":
-                index = -1
-                break
-            case "related":
-                // How do we simulate that?
-                break
-        }
-        // Then make sure the tab is made active and moved to the right place
-        // when it is opened in the current window
-        const selecttab = tab => {
-            browser.tabs.onCreated.removeListener(selecttab)
-            tabSetActive(tab.id)
-            browser.tabs.move(tab.id, { index })
-        }
-        browser.tabs.onCreated.addListener(selecttab)
-
-        try {
-            if ((await browser.runtime.getPlatformInfo()).os === "mac") {
-                if ((await browser.windows.getCurrent()).incognito) {
-                    throw "nativeopen isn't supported in private mode on OSX. Consider installing Linux or Windows :)."
-                }
-                const osascriptArgs = ["-e 'on run argv'", "-e 'tell application \"Firefox\" to open location item 1 of argv'", "-e 'end run'"]
-                await Native.run("osascript " + osascriptArgs.join(" ") + " " + url)
-            } else {
-                const os = (await browser.runtime.getPlatformInfo()).os
-                if (firefoxArgs.length === 0) {
-                    try {
-                        const profile = await Native.getProfile()
-                        if (profile.Name !== undefined) {
-                            firefoxArgs = [`-p ${profile.Name}`]
-                        } else if (profile.absolutePath !== undefined) {
-                            if (os === "win") {
-                                firefoxArgs = [`--profile "${profile.absolutePath}"`]
-                            } else {
-                                firefoxArgs = [`--profile '${profile.absolutePath}'`]
-                            }
-                        }
-                    } catch (e) {
-                        logger.debug(e)
-                        firefoxArgs = []
-                    }
-                    firefoxArgs.push("--new-tab")
-                }
-                let escapedUrl = url
-                // On linux, we need to quote and escape single quotes in the
-                // url, otherwise an attacker could create an anchor with a url
-                // like 'file:// && $(touch /tmp/dead)' and achieve remote code
-                // execution when the user tries to follow it with `hint -W tabopen`
-                // But windows treats single quotes as "open this file from the
-                // user's directory", so we need to use double quotes there
-                if (os === "win") {
-                    escapedUrl = `"${escapedUrl.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`
-                } else {
-                    escapedUrl = `'${escapedUrl.replace(/'/g, '"\'"')}'`
-                }
-                await Native.run(`${config.get("browser")} ${firefoxArgs.join(" ")} ${escapedUrl}`)
-            }
-            setTimeout(() => browser.tabs.onCreated.removeListener(selecttab), 100)
-        } catch (e) {
-            browser.tabs.onCreated.removeListener(selecttab)
-            throw e
-        }
-    }
+    return excmd_open.nativeopen(...args)
 }
 
 /**
@@ -758,12 +638,6 @@ export async function saveas(...filename: string[]) {
 }
 
 // }}}
-
-/** @hidden */
-//#background_helper
-function tabSetActive(id: number) {
-    return browser.tabs.update(id, { active: true })
-}
 
 // }}}
 
@@ -1074,11 +948,6 @@ export async function reloadallbut(hard = false) {
 export async function reloadhard(n = 1) {
     reload(n, true)
 }
-
-// I went through the whole list https://developer.mozilla.org/en-US/Firefox/The_about_protocol
-// about:blank is even more special
-/** @hidden */
-export const ABOUT_WHITELIST = ["about:license", "about:logo", "about:rights", "about:blank"]
 
 /**
  * Open a new page in the current tab.
@@ -1637,19 +1506,6 @@ export function urlmodify(mode: "-t" | "-r" | "-s" | "-q" | "-Q" | "-g", ...args
     }
 }
 
-/** Returns the url of links that have a matching rel.
-
-    Don't bind to this: it's an internal function.
-
-    @hidden
- */
-//#content
-export function geturlsforlinks(reltype = "rel", rel: string) {
-    const elems = document.querySelectorAll("link[" + reltype + "='" + rel + "']")
-    if (elems) return Array.prototype.map.call(elems, x => x.href)
-    return []
-}
-
 /** Sets the current page's zoom level anywhere between 30% and 300%.
  *
  * If you overshoot the level while using relative adjustments i.e. level > 300% or level < 30%
@@ -1880,27 +1736,13 @@ export function focusbyid(id: string) {
 
 // {{{ TABS
 
-/** Switch to the tab by index (position on tab bar), wrapping round.
-
-    @param index
-        1-based index of the tab to target. Wraps such that 0 = last tab, -1 =
-        penultimate tab, etc.
-
-        if undefined, return activeTabId()
-*/
-/** @hidden */
-//#background_helper
-async function tabIndexSetActive(index: number | string) {
-    return tabSetActive(await idFromIndex(index))
-}
-
 /** Switch to the next tab, wrapping round.
 
     If increment is specified, move that many tabs forwards.
  */
 //#background
 export async function tabnext(increment = 1) {
-    return tabprev(-increment)
+    return excmd_tabs.tabnext(increment)
 }
 
 /** Switch to the next tab, wrapping round.
@@ -1912,11 +1754,7 @@ export async function tabnext(increment = 1) {
  */
 //#background
 export async function tabnext_gt(index?: number) {
-    if (index === undefined) {
-        tabnext()
-    } else {
-        tabIndexSetActive(index)
-    }
+    return excmd_tabs.tabnext_gt(index)
 }
 
 /** Switch to the previous tab, wrapping round.
@@ -1925,14 +1763,7 @@ export async function tabnext_gt(index?: number) {
  */
 //#background
 export async function tabprev(increment = 1) {
-    // Proper way:
-    // return tabIndexSetActive((await activeTab()).index - increment + 1)
-    // Kludge until https://bugzilla.mozilla.org/show_bug.cgi?id=1504775 is fixed:
-    return browser.tabs.query({ currentWindow: true, hidden: false }).then(tabs => {
-        tabs.sort((t1, t2) => t1.index - t2.index)
-        const prevTab = (tabs.findIndex(t => t.active) - increment + tabs.length) % tabs.length
-        return browser.tabs.update(tabs[prevTab].id, { active: true })
-    })
+    return excmd_tabs.tabprev(increment)
 }
 
 /** Like [[open]], but in a new tab. If no address is given, it will open the newtab page, which can be set with `set newtab [url]`
@@ -1957,106 +1788,13 @@ export async function tabprev(increment = 1) {
 */
 //#background
 export async function tabopen(...addressarr: string[]) {
-    let active
-    let container
-
-    const win = await browser.windows.getCurrent()
-
-    // Lets us pass both -b and -c in no particular order as long as they are up front.
-    async function argParse(args): Promise<string[]> {
-        if (args[0] === "-b") {
-            active = false
-            args.shift()
-            argParse(args)
-        } else if (args[0] === "-c") {
-            // Ignore the -c flag if incognito as containers are disabled.
-            if (!win.incognito) container = await Container.fuzzyMatch(args[1])
-            else logger.error("[tabopen] can't open a container in a private browsing window.")
-
-            args.shift()
-            args.shift()
-            argParse(args)
-        }
-        return args
-    }
-
-    const query = await argParse(addressarr)
-
-    const address = query.join(" ")
-    if (!ABOUT_WHITELIST.includes(address) && address.match(/^(about|file):.*/)) {
-        return nativeopen(address)
-    }
-
-    const aucon = new AutoContain()
-    if (!container && aucon.autocontainConfigured()) {
-        const autoContainer = await aucon.getAuconForUrl(address)
-        if (autoContainer && autoContainer !== "firefox-default") {
-            container = autoContainer
-            logger.debug("tabopen setting container automatically using autocontain directive")
-        }
-    }
-
-    return activeTabContainerId().then(containerId => {
-        const args = { active } as any
-        // Ensure -c has priority.
-        if (container) {
-            args.cookieStoreId = container
-        } else if (containerId && config.get("tabopencontaineraware") === "true") {
-            args.cookieStoreId = containerId
-        }
-        return openInNewTab(null, args).then(tab => openInTab(tab, { loadReplace: true }, query))
-    })
-}
-
-/** Resolve a tab index to the tab id of the corresponding tab in this window.
-
-    @param index
-        1-based index of the tab to target. Wraps such that 0 = last tab, -1 =
-        penultimate tab, etc.
-
-        also supports # for previous tab, % for current tab.
-
-        if undefined, return activeTabId()
-
-    @hidden
-*/
-//#background_helper
-async function idFromIndex(index?: number | "%" | "#" | string): Promise<number> {
-    if (index === "#") {
-        // Support magic previous/current tab syntax everywhere
-        const tabs = await getSortedWinTabs()
-        if (tabs.length < 2) {
-            // In vim, '#' is the id of the previous buffer even if said buffer has been wiped
-            // However, firefox doesn't store tab ids for closed tabs
-            // Since vim makes '#' default to the current buffer if only one buffer has ever been opened for the current session, it seems reasonable to return the id of the current tab if only one tab is opened in firefox
-            return activeTabId()
-        }
-        return tabs[1].id
-    } else if (index !== undefined && index !== "%") {
-        // Wrap
-        index = Number(index)
-        index = (index - 1).mod((await browser.tabs.query({ currentWindow: true })).length) + 1
-
-        // Return id of tab with that index.
-        return (await browser.tabs.query({
-            currentWindow: true,
-            index: index - 1,
-        }))[0].id
-    } else {
-        return activeTabId()
-    }
+    return excmd_open.tabopen(...addressarr)
 }
 
 /** Close all other tabs in this window */
 //#background
 export async function tabonly() {
-    const tabs = await browser.tabs.query({
-        pinned: false,
-        active: false,
-        currentWindow: true,
-    })
-    const tabsIds = tabs.map(tab => tab.id)
-    browser.tabs.remove(tabsIds)
+    return excmd_tabs.tabonly()
 }
 
 /** Duplicate a tab.
@@ -2066,7 +1804,7 @@ export async function tabonly() {
 */
 //#background
 export async function tabduplicate(index?: number) {
-    browser.tabs.duplicate(await idFromIndex(index))
+    return excmd_tabs.tabduplicate(index)
 }
 
 /** Detach a tab, opening it in a new window.
@@ -2076,18 +1814,7 @@ export async function tabduplicate(index?: number) {
 */
 //#background
 export async function tabdetach(index?: number) {
-    browser.windows.create({ tabId: await idFromIndex(index) })
-}
-
-/** Get list of tabs sorted by most recent use
-
-    @hidden
-*/
-//#background_helper
-async function getSortedWinTabs(): Promise<browser.tabs.Tab[]> {
-    const tabs = await browser.tabs.query({ currentWindow: true })
-    tabs.sort((a, b) => (a.lastAccessed < b.lastAccessed ? 1 : -1))
-    return tabs
+    return excmd_tabs.tabdetach(index)
 }
 
 /** Toggle fullscreen state
@@ -2112,14 +1839,7 @@ export async function fullscreen() {
 */
 //#background
 export async function tabclose(...indexes: string[]) {
-    if (indexes.length > 0) {
-        let ids: number[]
-        ids = await Promise.all(indexes.map(index => idFromIndex(index)))
-        browser.tabs.remove(ids)
-    } else {
-        // Close current tab
-        browser.tabs.remove(await activeTabId())
-    }
+    return excmd_tabs.tabclose(...indexes)
 }
 
 /** Close all tabs to the right of the current one
@@ -2649,62 +2369,28 @@ export async function sleep(time_ms: number) {
     await new Promise(resolve => setTimeout(resolve, time_ms))
 }
 
-/** @hidden */
-//#content
-export function showcmdline(focus = true) {
-    CommandLineContent.show()
-    let done = Promise.resolve()
-    if (focus) {
-        CommandLineContent.focus()
-        done = Messaging.messageOwnTab("commandline_frame", "focus")
-    }
-    return done
-}
-
-/** @hidden */
-//#content
-export function hidecmdline() {
-    CommandLineContent.hide_and_blur()
-}
-
 /** Set the current value of the commandline to string *with* a trailing space */
 //#content
 export function fillcmdline(...strarr: string[]) {
-    const str = strarr.join(" ")
-    showcmdline()
-    return Messaging.messageOwnTab("commandline_frame", "fillcmdline", [str])
+    return excmd_fillcmdline.fillcmdline(...strarr)
 }
 
 /** Set the current value of the commandline to string *without* a trailing space */
 //#content
 export function fillcmdline_notrail(...strarr: string[]) {
-    const str = strarr.join(" ")
-    showcmdline()
-    return Messaging.messageOwnTab("commandline_frame", "fillcmdline", [str, false])
+    return excmd_fillcmdline.fillcmdline_notrail(...strarr)
 }
 
 /** Show and fill the command line without focusing it */
 //#content
 export function fillcmdline_nofocus(...strarr: string[]) {
-    showcmdline(false)
-    return Messaging.messageOwnTab("commandline_frame", "fillcmdline", [strarr.join(" "), false, false])
+    return excmd_fillcmdline.fillcmdline_nofocus(...strarr)
 }
 
 /** Shows str in the command line for ms milliseconds. Recommended duration: 3000ms. */
 //#content
 export async function fillcmdline_tmp(ms: number, ...strarr: string[]) {
-    const str = strarr.join(" ")
-    showcmdline(false)
-    Messaging.messageOwnTab("commandline_frame", "fillcmdline", [strarr.join(" "), false, false])
-    return new Promise(resolve =>
-        setTimeout(async () => {
-            if ((await Messaging.messageOwnTab("commandline_frame", "getContent", [])) === str) {
-                CommandLineContent.hide_and_blur()
-                resolve(Messaging.messageOwnTab("commandline_frame", "clear", [true]))
-            }
-            resolve()
-        }, ms),
-    )
+    return excmd_fillcmdline.fillcmdline_tmp(ms, ...strarr)
 }
 
 /**
@@ -2716,54 +2402,21 @@ export async function get_current_url() {
 }
 
 /**
- * Copy content to clipboard without feedback. Use `clipboard yank` for interactive use.
- */
-//#background
-export function yank(...content: string[]) {
-    return setclip(content.join(" "))
-}
-
-/**
- * Copies a string to the clipboard/selection buffer depending on the user's preferences
- *
- * @hidden
- */
-//#background_helper
-async function setclip(str) {
-    // Functions to avoid retyping everything everywhere
-
-    // Note: We're using fillcmdline here because exceptions are somehow not caught. We're rethrowing because otherwise the error message will be overwritten with the "yank successful" message.
-    const s = () => Native.clipboard("set", str)
-    const c = () => messageActiveTab("commandline_frame", "setClipboard", [str])
-
-    let promises = []
-    switch (await config.getAsync("yankto")) {
-        case "selection":
-            promises = [s()]
-            break
-        case "clipboard":
-            promises = [c()]
-            break
-        case "both":
-            promises = [s(), c()]
-            break
-    }
-    return Promise.all(promises)
-}
-
-/**
  * Fetches the content of the clipboard/selection buffer depending on user's preferences
  *
  * Exposed for use with [[composite]], e.g. `composite getclip | fillcmdline`
  */
 //#background
 export async function getclip(fromm?: "clipboard" | "selection") {
-    if (fromm === undefined) fromm = await config.getAsync("putfrom")
-    if (fromm === "clipboard") {
-        return messageActiveTab("commandline_frame", "getClipboard")
-    } else {
-        return Native.clipboard("get", "")
-    }
+    return excmd_clipboard.getclip(fromm)
+}
+
+/**
+ * Copy content to clipboard without feedback. Use `clipboard yank` for interactive use.
+ */
+//#background
+export function yank(...content: string[]) {
+    return excmd_clipboard.yank(...content)
 }
 
 /** Use the system clipboard.
@@ -2789,66 +2442,7 @@ export async function getclip(fromm?: "clipboard" | "selection") {
 */
 //#background
 export async function clipboard(excmd: "open" | "yank" | "yankshort" | "yankcanon" | "yanktitle" | "yankmd" | "xselpaste" | "tabopen" = "open", ...toYank: string[]) {
-    let content = toYank.join(" ")
-    let url = ""
-    let urls = []
-    switch (excmd) {
-        case "yankshort":
-            urls = await geturlsforlinks("rel", "shortlink")
-            if (urls.length === 0) {
-                urls = await geturlsforlinks("rev", "canonical")
-            }
-            if (urls.length > 0) {
-                await yank(urls[0])
-                fillcmdline_tmp(3000, "# " + urls[0] + " copied to clipboard.")
-                break
-            }
-        // Trying yankcanon if yankshort failed...
-        case "yankcanon":
-            urls = await geturlsforlinks("rel", "canonical")
-            if (urls.length > 0) {
-                await yank(urls[0])
-                fillcmdline_tmp(3000, "# " + urls[0] + " copied to clipboard.")
-                break
-            }
-        // Trying yank if yankcanon failed...
-        case "yank":
-            content = content === "" ? (await activeTab()).url : content
-            await yank(content)
-            fillcmdline_tmp(3000, "# " + content + " copied to clipboard.")
-            break
-        case "yanktitle":
-            content = (await activeTab()).title
-            await yank(content)
-            fillcmdline_tmp(3000, "# " + content + " copied to clipboard.")
-            break
-        case "yankmd":
-            content = "[" + (await activeTab()).title + "](" + (await activeTab()).url + ")"
-            await yank(content)
-            fillcmdline_tmp(3000, "# " + content + " copied to clipboard.")
-            break
-        case "open":
-            url = await getclip()
-            if (url) {
-                open(url)
-            }
-            break
-        case "tabopen":
-            url = await getclip()
-            if (url) {
-                tabopen(url)
-            }
-            break
-        case "xselpaste":
-            content = await getclip("selection")
-            if (content.length > 0) {
-                EditorCmds.insert_text(content)
-            }
-            break
-        default:
-            // todo: maybe we should have some common error and error handler
-            throw new Error(`[clipboard] unknown excmd: ${excmd}`)
-    }
+    return excmd_clipboard.clipboard(excmd, ...toYank)
 }
 
 /** Change active tab.
@@ -2862,7 +2456,7 @@ export async function clipboard(excmd: "open" | "yank" | "yankshort" | "yankcano
  */
 //#background
 export async function tab(index: number | "#") {
-    tabIndexSetActive(index)
+    return excmd_tabs.tab(index)
 }
 
 /** Change active tab.
@@ -2873,16 +2467,7 @@ export async function tab(index: number | "#") {
  */
 //#background
 export async function taball(id: string) {
-    const windows = (await browser.windows.getAll()).map(w => w.id).sort((a, b) => a - b)
-    if (id === null || id === undefined || !id.match(/\d+\.\d+/)) {
-        const tab = await activeTab()
-        const prevId = id
-        id = windows.indexOf(tab.windowId) + "." + (tab.index + 1)
-        logger.info(`taball: Bad tab id: ${prevId}, defaulting to ${id}`)
-    }
-    const [winindex, tabindex] = id.split(".")
-    await browser.windows.update(windows[parseInt(winindex, 10) - 1], { focused: true })
-    return browser.tabs.update(await idFromIndex(tabindex), { active: true })
+    return excmd_tabs.taball(id)
 }
 
 // }}}
